修复 Qt 信号槽类型不匹配：interactionStateChanged 使用 int 类型
## 编译错误

```
error: static assertion failed: Signal and slot arguments are not compatible.
```

**原因**: Qt 的信号槽系统不支持枚举类型到 int 的自动转换

**位置**:
- 信号: `ChartView::interactionStateChanged(InteractionState)`
- 槽: `lambda(int newState)`

## 修复方案

### 修改信号定义
```cpp
// ❌ 之前：发出枚举类型
void interactionStateChanged(InteractionState newState);

// ✅ 现在：发出 int 类型
void interactionStateChanged(int newState);
```

### 修改信号发送
```cpp
// ❌ 之前：直接发出枚举
emit interactionStateChanged(m_interactionState);

// ✅ 现在：转换为 int
emit interactionStateChanged(static_cast<int>(m_interactionState));
```

## 修改位置

**chart_view.h**:
- 第 153 行: 修改信号声明为 `int` 类型
- 添加注释说明枚举值对应关系 (0=Idle, 1=WaitingForPoints, 2=PointsCompleted, 3=Executing)

**chart_view.cpp**:
- 第 163 行: `handlePointSelectionClick()` 中发送信号
- 第 304 行: `startAlgorithmInteraction()` 中发送信号
- 第 327 行: `cancelAlgorithmInteraction()` 中发送信号

## 技术说明

### Qt 元对象系统限制

Qt 的元对象系统在运行时检查信号和槽的参数类型。虽然 C++ 允许枚举到整数的隐式转换，但 Qt 的元对象系统要求类型完全匹配或者显式注册转换。

### 替代方案（未采用）

1. **使用 Q_DECLARE_METATYPE**:
   ```cpp
   Q_DECLARE_METATYPE(ChartView::InteractionState)
   ```
   - 需要在多个地方注册类型
   - 增加编译依赖

2. **使用 qRegisterMetaType**:
   ```cpp
   qRegisterMetaType<ChartView::InteractionState>("ChartView::InteractionState");
   ```
   - 需要在初始化时注册
   - 运行时开销

3. **匹配枚举类型**:
   ```cpp
   [](ChartView::InteractionState newState) { ... }
   ```
   - 需要包含 chart_view.h
   - 增加编译依赖

### 选择的方案优势

✅ **简单直接**: 不需要额外的类型注册
✅ **减少依赖**: lambda 不需要包含 chart_view.h
✅ **性能最优**: 无运行时类型转换开销
✅ **向后兼容**: 枚举值可以直接当 int 使用

## 验证

编译应通过以下检查:
- ✅ 信号和槽参数类型匹配 (int == int)
- ✅ 无需额外的类型注册
- ✅ 功能保持不变

相关提交:
- ce50853: 实现活动算法状态机
- bc68db8: MainController 集成
- 02269c1: 移除向后兼容代码
- 77a3fd5: 修复编译错误（第一轮）

+++ Analysis/src/ui/chart_view.cpp
@@ -160,7 +160,7 @@ void ChartView::handlePointSelectionClick(const QPointF& chartViewPos)
     if (m_selectedPoints.size() >= m_activeAlgorithm.requiredPointCount) {
         // 状态转换: WaitingForPoints → PointsCompleted
         m_interactionState = InteractionState::PointsCompleted;
-        emit interactionStateChanged(m_interactionState);
+        emit interactionStateChanged(static_cast<int>(m_interactionState));
 
         qDebug() << "ChartView: 算法" << m_activeAlgorithm.displayName
                  << "交互完成，发送信号触发执行";
@@ -301,7 +301,7 @@ void ChartView::startAlgorithmInteraction(const QString& algorithmName, const QS
 
     // 状态转换: Idle → WaitingForPoints
     m_interactionState = InteractionState::WaitingForPoints;
-    emit interactionStateChanged(m_interactionState);
+    emit interactionStateChanged(static_cast<int>(m_interactionState));
 
     // 切换到选点模式
     setInteractionMode(InteractionMode::Pick);
@@ -324,7 +324,7 @@ void ChartView::cancelAlgorithmInteraction()
 
     // 状态转换: 任意状态 → Idle
     m_interactionState = InteractionState::Idle;
-    emit interactionStateChanged(m_interactionState);
+    emit interactionStateChanged(static_cast<int>(m_interactionState));
 
     // 切换回视图模式
     setInteractionMode(InteractionMode::View);
